+++ tests/core/test_model.py
@@ -32,7 +32,7 @@ import flash
 from flash.core.adapter import Adapter
 from flash.core.classification import ClassificationTask
 from flash.core.data.process import DefaultPreprocess, Postprocess
-from flash.core.utilities.imports import _PIL_AVAILABLE, _TABULAR_AVAILABLE, _TEXT_AVAILABLE
+from flash.core.utilities.imports import _TABULAR_AVAILABLE, _TEXT_AVAILABLE, Image
 from flash.image import ImageClassificationData, ImageClassifier
 from tests.helpers.utils import _IMAGE_TESTING, _TABULAR_TESTING
 
@@ -40,24 +40,6 @@ if _TABULAR_AVAILABLE:
     from flash.tabular import TabularClassifier
 else:
     TabularClassifier = None
-
-if _PIL_AVAILABLE:
-    from PIL import Image
-else:
-
-    class MetaImage(type):
-        def __init__(cls, name, bases, dct):
-            super(MetaImage, cls).__init__(name, bases, dct)
-
-            cls._Image = None
-
-        @property
-        def Image(cls):
-            warn("Mock object called due to missing PIL library. Install PIL using 'pip install Pillow'.")
-            return cls._Image
-
-    class Image(metaclass=MetaImage):
-        pass
 
 
 # ======== Mock functions ========
